+++ euphoria-service/src/main/java/club/euphoria_recruiting/handler/CompanyHandlers.java
@@ -1,61 +1,64 @@
-package edu.cooper.ece366.euphoria.handler;
+package club.euphoria_recruiting.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import com.spotify.apollo.RequestContext;
 import com.spotify.apollo.Response;
 import com.spotify.apollo.route.*;
-import edu.cooper.ece366.euphoria.model.Cookie;
-import edu.cooper.ece366.euphoria.store.model.CookieStore;
+import edu.cooper.ece366.euphoria.model.Company;
+import edu.cooper.ece366.euphoria.store.model.CompanyStore;
 import okio.ByteString;
 
 import java.io.IOException;
 import java.util.Map;
 import java.util.stream.Stream;
 
-public class CookieHandlers implements RouteProvider {
+public class CompanyHandlers implements RouteProvider {
     private final ObjectMapper objectMapper;
-    private final CookieStore cookieStore;
+    private final CompanyStore companyStore;
 
-    public CookieHandlers(final ObjectMapper objectMapper, CookieStore cookieStore) {
+    public CompanyHandlers(final ObjectMapper objectMapper, CompanyStore companyStore) {
         this.objectMapper = objectMapper;
-        this.cookieStore = cookieStore;
+        this.companyStore = companyStore;
     }
 
     @Override
     public Stream<Route<AsyncHandler<Response<ByteString>>>> routes() {
         return Stream.of(
-                Route.sync("GET", "/api/cookie/<cookieCheck>", this::getCookie),
-                Route.sync("POST", "/api/cookie", this::createCookie)
+                Route.sync("GET", "/api/company/<companyId>", this::getCompany),
+                Route.sync("POST", "/api/company", this::createCompany)
         ).map(r -> r.withMiddleware(jsonMiddleware()));
     }
 
     @VisibleForTesting
-    public Cookie getCookie(final RequestContext rc) {
-        return cookieStore.getCookie(rc.pathArgs().get("cookieCheck"));
+    public Company getCompany(final RequestContext rc) {
+        return companyStore.getCompany(rc.pathArgs().get("companyId"));
     }
 
     @VisibleForTesting
-    public Cookie createCookie(final RequestContext rc) {
-        String username = null;
-        String passwordHash = null;
+    public Company createCompany(final RequestContext rc) {
+        String name, website, description;
+        name = null;
+        website = null;
+        description = null;
         boolean success = false;
         try {
-            Map jsonMap = objectMapper.readValue(rc.request().payload().get().toByteArray(), Map.class);
-            username = jsonMap.get("username").toString();
-            passwordHash = jsonMap.get("passwordHash").toString();
+            byte[] requestBytes = rc.request().payload().get().toByteArray();
+            Map jsonMap = objectMapper.readValue(requestBytes, Map.class);
+            name = jsonMap.get("name").toString();
+            website = jsonMap.get("website").toString();
+            description = jsonMap.get("description").toString();
             success = true;
         } catch (IOException ex) {
             System.out.println(ex);
         }
 
         if (success)
-            return cookieStore.createCookie(username, passwordHash);
+            return companyStore.createCompany(name, website, description);
         else
             return null;
 
     }
-
 
     private <T> Middleware<AsyncHandler<T>, AsyncHandler<Response<ByteString>>> jsonMiddleware() {
         return JsonSerializerMiddlewares.<T>jsonSerialize(objectMapper.writer())
