+++ euphoria-service/src/main/java/club/euphoria_recruiting/handler/CookieHandlers.java
@@ -1,64 +1,61 @@
-package edu.cooper.ece366.euphoria.handler;
+package club.euphoria_recruiting.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import com.spotify.apollo.RequestContext;
 import com.spotify.apollo.Response;
 import com.spotify.apollo.route.*;
-import edu.cooper.ece366.euphoria.model.Company;
-import edu.cooper.ece366.euphoria.store.model.CompanyStore;
+import edu.cooper.ece366.euphoria.model.Cookie;
+import edu.cooper.ece366.euphoria.store.model.CookieStore;
 import okio.ByteString;
 
 import java.io.IOException;
 import java.util.Map;
 import java.util.stream.Stream;
 
-public class CompanyHandlers implements RouteProvider {
+public class CookieHandlers implements RouteProvider {
     private final ObjectMapper objectMapper;
-    private final CompanyStore companyStore;
+    private final CookieStore cookieStore;
 
-    public CompanyHandlers(final ObjectMapper objectMapper, CompanyStore companyStore) {
+    public CookieHandlers(final ObjectMapper objectMapper, CookieStore cookieStore) {
         this.objectMapper = objectMapper;
-        this.companyStore = companyStore;
+        this.cookieStore = cookieStore;
     }
 
     @Override
     public Stream<Route<AsyncHandler<Response<ByteString>>>> routes() {
         return Stream.of(
-                Route.sync("GET", "/api/company/<companyId>", this::getCompany),
-                Route.sync("POST", "/api/company", this::createCompany)
+                Route.sync("GET", "/api/cookie/<cookieCheck>", this::getCookie),
+                Route.sync("POST", "/api/cookie", this::createCookie)
         ).map(r -> r.withMiddleware(jsonMiddleware()));
     }
 
     @VisibleForTesting
-    public Company getCompany(final RequestContext rc) {
-        return companyStore.getCompany(rc.pathArgs().get("companyId"));
+    public Cookie getCookie(final RequestContext rc) {
+        return cookieStore.getCookie(rc.pathArgs().get("cookieCheck"));
     }
 
     @VisibleForTesting
-    public Company createCompany(final RequestContext rc) {
-        String name, website, description;
-        name = null;
-        website = null;
-        description = null;
+    public Cookie createCookie(final RequestContext rc) {
+        String username = null;
+        String passwordHash = null;
         boolean success = false;
         try {
-            byte[] requestBytes = rc.request().payload().get().toByteArray();
-            Map jsonMap = objectMapper.readValue(requestBytes, Map.class);
-            name = jsonMap.get("name").toString();
-            website = jsonMap.get("website").toString();
-            description = jsonMap.get("description").toString();
+            Map jsonMap = objectMapper.readValue(rc.request().payload().get().toByteArray(), Map.class);
+            username = jsonMap.get("username").toString();
+            passwordHash = jsonMap.get("passwordHash").toString();
             success = true;
         } catch (IOException ex) {
             System.out.println(ex);
         }
 
         if (success)
-            return companyStore.createCompany(name, website, description);
+            return cookieStore.createCookie(username, passwordHash);
         else
             return null;
 
     }
+
 
     private <T> Middleware<AsyncHandler<T>, AsyncHandler<Response<ByteString>>> jsonMiddleware() {
         return JsonSerializerMiddlewares.<T>jsonSerialize(objectMapper.writer())
