+++ euphoria-service/src/main/java/club/euphoria_recruiting/model/Cookie.java
