+++ euphoria-service/src/main/java/club/euphoria_recruiting/model/Posting.java
