+++ euphoria-service/src/main/java/club/euphoria_recruiting/store/jdbc/ApplicationStoreJdbc.java
