+++ euphoria-service/src/main/java/club/euphoria_recruiting/store/model/UserStore.java
