+++ euphoria-service/src/main/java/club/euphoria_recruiting/utils/Industry.java
